+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Enhanced: Variable validation by removing parameters not used.
 * Fixed: Undefined class constant `E_WC_ADMIN_NOTE_MARKETING` for those who are using WooCommerce lower than version `4.3.0`. Thanks to [@dmchale](https://github.com/dmchale)
 * Fixed: If stock not available when updating item, return error. Thanks to [@pauld8](https://github.com/pauld8)
+* Fixed: Product ID validation should the ID not be present. Also returns correct product ID should SKU ID be used instead.
 * Tested: Compatible with WooCommerce v4.6.0
 
 ## v2.6.3 - 23rd September, 2020